Move snapshot stuff to org.rust.lang.utils.snapshot
