+++ src/main/kotlin/org/rust/lang/utils/snapshot/Snapshot.kt
@@ -3,11 +3,10 @@
  * found in the LICENSE file.
  */
 
-package org.rust.lang.core.types.infer
+package org.rust.lang.utils.snapshot
 
 interface Snapshot {
     fun commit()
-
     fun rollback()
 }
 
@@ -30,7 +29,7 @@ class UndoLog {
 private class LogBasedSnapshot private constructor(
     private val undoLog: MutableList<Undoable>,
     val position: Int
-): Snapshot {
+) : Snapshot {
     override fun commit() {
         assertOpenSnapshot()
         if (position == 0) {
@@ -40,7 +39,7 @@ private class LogBasedSnapshot private constructor(
         }
     }
 
-    override fun rollback(){
+    override fun rollback() {
         assertOpenSnapshot()
         val toRollback = undoLog.subList(position + 1, undoLog.size)
         toRollback.asReversed().forEach(Undoable::undo)
@@ -74,4 +73,9 @@ private class LogBasedSnapshot private constructor(
             // the inner is committed but outer is rolled back.
         }
     }
+}
+
+class CombinedSnapshot(vararg val snapshots: Snapshot) : Snapshot {
+    override fun rollback() = snapshots.forEach { it.rollback() }
+    override fun commit() = snapshots.forEach { it.commit() }
 }
